Command/Walkthrough logger dichotomy
