+++ src/main/java/fr/estecka/variantscit/modules/IBakedModule.java
@@ -5,6 +5,7 @@ import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.item.ItemStack;
 import fr.estecka.variantscit.VariantsCitMod;
 import fr.estecka.variantscit.commands.CommandLogger;
+import fr.estecka.variantscit.commands.WalktroughLogger;
 import fr.estecka.variantscit.modules.cache.ICacheKey;
 
 public interface IBakedModule
@@ -40,7 +41,7 @@ extends ICacheKey.Cacheable
 		logger.Error("This module type does not support `dump`. Please report this issue.");
 	}
 
-	default ResourceLocation Walkthrough(CommandLogger logger, ItemStack stack) {
+	default ResourceLocation Walkthrough(WalktroughLogger logger, ItemStack stack) {
 		logger.Error("This module type does not support `walkthrough`. Please report this issue.");
 		return this.GetModelForItem(stack);
 	}
